chore: optimize static routes and metadata handling for Next.js App Router
