+++ src/app/about/metadata.ts
@@ -1,0 +1,5 @@
+export const dynamic = 'force-static';
+export const metadata = {
+  title: 'About | Luis Pulido',
+  description: 'Learn about Luis Pulido, a software engineer with over 12 years of experience building scalable solutions',
+};
