+++ src/app/about/page.tsx
@@ -1,9 +1,6 @@
-import Link from 'next/link';
 
-export const metadata = {
-  title: 'About | Luis Pulido',
-  description: 'Learn about Luis Pulido, a software engineer with over 12 years of experience building scalable solutions',
-};
+import { dynamic, metadata } from './metadata';
+import Link from 'next/link';
 
 export default function About() {
   return (
