+++ src/app/contact/metadata.ts
@@ -1,0 +1,5 @@
+export const dynamic = 'force-static';
+export const metadata = {
+  title: 'Contact | Luis Pulido',
+  description: 'Get in touch with Luis Pulido',
+};
