+++ src/app/page.tsx
@@ -1,3 +1,4 @@
+export const dynamic = 'force-static';
 import { getSortedPostsData } from '@/lib/posts';
 import Link from 'next/link';
 import { Suspense } from 'react';
