+++ src/app/random/metadata.ts
@@ -1,0 +1,9 @@
+export const dynamic = 'force-static';
+export const metadata = {
+  title: "Random | Luis Pulido Díaz",
+  description: "Random creative experiments and interactive visualizations",
+  openGraph: {
+    title: "Random | Luis Pulido Díaz",
+    description: "Random creative experiments and interactive visualizations",
+  },
+};
