+++ src/app/random/smudge-canvas/page.tsx
@@ -1,4 +1,6 @@
+
 "use client";
+export const dynamic = 'force-static';
 
 import SmudgeCanvas from "@/components/SmudgeCanvas";
 import Link from "next/link";
