+++ src/app/terms-of-service/page.tsx
@@ -1,3 +1,5 @@
+"use client";
+export const dynamic = 'force-static';
 export default function TermsOfServicePage() {
   return (
     <main className="prose mx-auto p-8">
